Perform whole-program validation for TypeScript code, too.
