+++ package.json
@@ -32,10 +32,11 @@
   "dependencies": {},
   "devDependencies": {
     "babel-types": "^6.26.0",
-    "babylon": "^7.0.0-beta.31",
+    "babylon": "^7.0.0-beta.40",
     "espree": "^3.1.7",
     "esprima": "~4.0.0",
     "esprima-fb": "~14001.1.0-dev-harmony-fb",
+    "glob": "^7.1.2",
     "mocha": "^5.0.0",
     "reify": "^0.14.1"
   },
